+++ Rules/IPv4/Outbound/WindowsServices.ps1
@@ -222,6 +222,11 @@ if ($Service)
 local and push notifications. Supported notifications are tile, toast and raw." `
 		@Logs | Format-Output @Logs
 }
+else
+{
+	# NOTE: this may not be found on fresh installed system, need to reload later
+	Write-Warning -Message "Windows Push Notifications User Service was not found"
+}
 
 New-NetFirewallRule -DisplayName "Windows Insider Service" `
 	-Platform $Platform -PolicyStore $PolicyStore -Profile $FirewallProfile `